+++ api/include/opentelemetry/trace/propagation/text_map_propagator.h
@@ -11,32 +11,32 @@ namespace trace
 namespace propagation
 {
 
+// TextMapCarrier is the storage medium used by TextMapPropagator.
+class TextMapCarrier
+{
+public:
+  /*returns the value associated with the passed key.*/
+  virtual nostd::string_view Get(nostd::string_view key) const noexcept = 0;
+
+  /*stores the key-value pair.*/
+  virtual void Set(nostd::string_view key, nostd::string_view value) noexcept = 0;
+};
+
 // The TextMapPropagator class provides an interface that enables extracting and injecting
 // context into carriers that travel in-band across process boundaries. HTTP frameworks and clients
 // can integrate with TextMapPropagator by providing the object containing the
 // headers, and a getter and setter function for the extraction and
 // injection of values, respectively.
-template <typename T>
+
 class TextMapPropagator
 {
 public:
-  // Rules that manages how context will be extracted from carrier.
-  using Getter = nostd::string_view (*)(const T &carrier, nostd::string_view trace_type);
-
-  // Rules that manages how context will be injected to carrier.
-  using Setter = void (*)(T &carrier,
-                          nostd::string_view trace_type,
-                          nostd::string_view trace_description);
-
-  // Returns the context that is stored in the carrier with the getter as extractor.
-  virtual context::Context Extract(Getter get_from_carrier,
-                                   const T &carrier,
+  // Returns the context that is stored in the carrier with the TextMapCarrier as extractor.
+  virtual context::Context Extract(const TextMapCarrier &carrier,
                                    context::Context &context) noexcept = 0;
 
   // Sets the context for carrier with self defined rules.
-  virtual void Inject(Setter set_from_carrier,
-                      T &carrier,
-                      const context::Context &context) noexcept = 0;
+  virtual void Inject(TextMapCarrier &carrier, const context::Context &context) noexcept = 0;
 };
 }  // namespace propagation
 }  // namespace trace