+++ exporters/otlp/include/opentelemetry/exporters/otlp/recordable.h
@@ -29,7 +29,7 @@ public:
                     const opentelemetry::common::AttributeValue &value) noexcept override;
 
   void AddEvent(nostd::string_view name,
-                core::SystemTimestamp timestamp,
+                common::SystemTimestamp timestamp,
                 const common::KeyValueIterable &attributes) noexcept override;
 
   void AddLink(const opentelemetry::trace::SpanContext &span_context,
@@ -41,9 +41,10 @@ public:
 
   void SetSpanKind(opentelemetry::trace::SpanKind span_kind) noexcept override;
 
+
   void SetResource(const opentelemetry::sdk::resource::Resource &resource) noexcept override;
 
-  void SetStartTime(opentelemetry::core::SystemTimestamp start_time) noexcept override;
+  void SetStartTime(opentelemetry::common::SystemTimestamp start_time) noexcept override;
 
   void SetDuration(std::chrono::nanoseconds duration) noexcept override;