+++ sdk/src/trace/tracer_context.cc
@@ -9,18 +9,27 @@ namespace trace
 
 TracerContext::TracerContext(std::unique_ptr<SpanProcessor> processor,
                              opentelemetry::sdk::resource::Resource resource,
-                             std::unique_ptr<Sampler> sampler) noexcept
-    : processor_(std::move(processor)), resource_(resource), sampler_(std::move(sampler))
+                             std::unique_ptr<Sampler> sampler,
+                             std::unique_ptr<IdGenerator> id_generator) noexcept
+    : processor_(std::move(processor)),
+      resource_(resource),
+      sampler_(std::move(sampler)),
+      id_generator_(std::move(id_generator))
 {}
 
 Sampler &TracerContext::GetSampler() const noexcept
 {
-  return *sampler_.get();
+  return *sampler_;
 }
 
 const opentelemetry::sdk::resource::Resource &TracerContext::GetResource() const noexcept
 {
   return resource_;
+}
+
+opentelemetry::sdk::trace::IdGenerator &TracerContext::GetIdGenerator() const noexcept
+{
+  return *id_generator_;
 }
 
 void TracerContext::RegisterPipeline(std::unique_ptr<SpanProcessor> processor) noexcept