+++ sdk/test/trace/span_data_test.cc
@@ -40,7 +40,7 @@ TEST(SpanData, Set)
       trace_id, span_id,
       opentelemetry::trace::TraceFlags{opentelemetry::trace::TraceFlags::kIsSampled}, true,
       trace_state};
-  opentelemetry::core::SystemTimestamp now(std::chrono::system_clock::now());
+  opentelemetry::common::SystemTimestamp now(std::chrono::system_clock::now());
 
   SpanData data;
   data.SetIdentity(span_context, parent_span_id);